[rom] Relocate exception handlers to their own section
1. Move the C-based ISR functions to their own module to better control
   the link order.
